Add lua graphical line chart plotting feature
Add lua chart plotting feature which allows to sample data from
instruments and plot the data in a nice graphical chart live!

This is a very powerful and flexible way to depict data and it is fast
because GTK4 utilizes GPU HW acceleration for rendering the chart.

Future planned features include the possibility to dump both the image
and the raw RSV data of a chart via GUI or via script. Of course, at
some point other chart types will be added.