Remove DS1074Z+ note
We do not want to start listing bug notes for specific instruments
unless they are directly related to a bug or limitation of lxi-tools -
we don't want to clutter the README. Technically the instrument works.
If people run into this issue they will most likely find the github
issue mentioning it.