Add support for manually adding instruments
Manually added instruments are managed via settings in the dconf registry
so that they can be remembered by the application.

Manually added instruments will appear with an instrument icon with a
yellow indicator (indicating that it may or may not be present).

Automatically discovered instruments will appear with an instrument icon
with a green indicator (indicating that it is likely present since it
responded).

Once added it is possible to edit or remove instruments again via the
right click menu.